refactor(frontend): split UserFileStatus into fileStatus/* components
Phase 5.4 — no-behavior/no-visual split (2051 -> 717). Both call sites (gallery /,
/file-status) still render <UserFileStatus/> with no props:
- fileStatus/TaskFilterPanel.svelte (359) — quick-filter chips + compact filter toolbar +
  inline date picker; two-way bind: filters so the parent's reactive refetch fires as before.
- fileStatus/TasksGrid.svelte (360) — loading/error/empty + tasks table + SearchPagination;
  events viewDetails/pageChange.
- fileStatus/FileDetailModal.svelte (534) — file-details modal; events close/retry (scroll-lock
  + retry orchestration stay in parent).
UserFileStatus keeps data loading, the WebSocket subscription + update handling, auto-refresh,
pagination/retry/recovery orchestration, status cards, and layout.

Also removed ~20 pre-existing DEAD CSS selectors (problems/recent-section toggles consolidated
away earlier; zero matching markup) that Svelte's unused-selector check surfaced once the
modal's scoped CSS moved out — no visual/behavior change.

Verified: svelte-check 0/0 (864 files), build, 52 unit tests, and 9 E2E green (settings 3/3 +
file-detail 6/6 — app loads with UserFileStatus on /).
